perf: "사진 업로드 로직 분리 및 변경과 타입 분리"
